ECA-8672: Remove unnecessary boxing/unboxing. Fix some variable names
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@33806 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/cesecore-common/src/org/cesecore/certificates/util/DnComponents.java
@@ -335,7 +335,7 @@ public class DnComponents {
             log.error("No dn id mapping from profile id " + profileId);
             // We allow it to fail here
         }
-        return val.intValue();
+        return val;
     }
 
     /**
